StackRouter.getStateForAction: replace instead of push when only child changes
This is a fix for this issue: https://github.com/react-community/react-navigation/issues/1508

It's more of a preliminary proposal, as I'm not yet sure all the far-reaching effects. What needs to be tested is a DrawerRouter and another StackRouter as a the child. This assumes the TabRouter is a child and works nicely. I'll hopefully be testing other routers soon and pushing any changes. If anyone with more experience with `react-navigation` could take a look at this, I'd greatly appreciate it.