fix($StackRouter): prevent pushing duplicate routes even without children
Basically the same sort of issue exists even without a child route. You end up in situations where you slide over a TabNavigator, and immediately push on to it instead of simply show what's already there. Probably throughout the codebase prevention of pushing duplicates should be blocked. If the `routeName` and `params` are the same, the screen is the same, which results in subpart user-experience in places where it's not an issue, and as just described there are some places this absolutely doesn't make sense. Perhaps the develop should just be `console.warned` when pushing duplicates so they know they're doing something wrong/unnecessary, and then in userland can change their code so as not to have buttons that don't do anything.